Create LatchAsync client
